+++ python/GafferSceneTest/ShaderTweakProxyTest.py
@@ -57,8 +57,8 @@ class ShaderTweakProxyTest( GafferSceneTest.SceneTestCase ) :
 
 		textureShader2 = GafferSceneTest.TestShader( "texture2" )
 
-		shader["parameters"]["c"].setInput( textureShader1["out"] )
-		textureShader1["parameters"]["c"].setInput( textureShader2["out"] )
+		shader["parameters"]["c"].setInput( textureShader1["out"]["c"] )
+		textureShader1["parameters"]["c"].setInput( textureShader2["out"]["c"] )
 
 		planeFilter = GafferScene.PathFilter()
 		planeFilter["paths"].setValue( IECore.StringVectorData( [ "/plane" ] ) )
@@ -71,7 +71,7 @@ class ShaderTweakProxyTest( GafferSceneTest.SceneTestCase ) :
 		# Check the untweaked network
 		originalNetwork = assignment["out"].attributes( "/plane" )["surface"]
 		self.assertEqual( len( originalNetwork ), 3 )
-		self.assertEqual( originalNetwork.input( ( "surface", "c" ) ), ( "texture1", "out" ) )
+		self.assertEqual( originalNetwork.input( ( "surface", "c" ) ), ( "texture1", "c" ) )
 
 		tweakShader = GafferSceneTest.TestShader( "tweakShader" )
 
@@ -81,12 +81,12 @@ class ShaderTweakProxyTest( GafferSceneTest.SceneTestCase ) :
 		tweaks["shader"].setValue( "surface" )
 
 		tweaks["tweaks"].addChild( Gaffer.TweakPlug( "c", Gaffer.Color3fPlug() ) )
-		tweaks["tweaks"][0]["value"].setInput( tweakShader["out"] )
+		tweaks["tweaks"][0]["value"].setInput( tweakShader["out"]["c"] )
 
 		# If we replace the upstream network with a tweak, now we have just 2 nodes
 		tweakedNetwork = tweaks["out"].attributes( "/plane" )["surface"]
 		self.assertEqual( len( tweakedNetwork ), 2 )
-		self.assertEqual( tweakedNetwork.input( ( "surface", "c" ) ), ( "tweakShader", "out" ) )
+		self.assertEqual( tweakedNetwork.input( ( "surface", "c" ) ), ( "tweakShader", "c" ) )
 
 		autoProxy = GafferScene.ShaderTweakProxy()
 		autoProxy.setupAutoProxy( Gaffer.Color3fPlug() )
@@ -97,20 +97,20 @@ class ShaderTweakProxyTest( GafferSceneTest.SceneTestCase ) :
 
 		# Test adding a tweak shader in the middle of the network using the proxy
 		tweakShader["parameters"]["c"].setInput( autoProxy["out"]["auto"] )
-		tweaks["tweaks"][0]["value"].setInput( tweakShader["out"] )
+		tweaks["tweaks"][0]["value"].setInput( tweakShader["out"]["c"] )
 		tweakedNetwork = tweaks["out"].attributes( "/plane" )["surface"]
 		self.assertEqual( len( tweakedNetwork ), 4 )
-		self.assertEqual( tweakedNetwork.input( ( "surface", "c" ) ), ( "tweakShader", "out" ) )
-		self.assertEqual( tweakedNetwork.input( ( "tweakShader", "c" ) ), ( "texture1", "out" ) )
+		self.assertEqual( tweakedNetwork.input( ( "surface", "c" ) ), ( "tweakShader", "c" ) )
+		self.assertEqual( tweakedNetwork.input( ( "tweakShader", "c" ) ), ( "texture1", "c" ) )
 
 		# If we target the end of the network where there is no input, then the tweak gets inserted fine,
 		# and there is no input to the tweak, since there's nothing upstream
 		tweaks["tweaks"][0]["name"].setValue( "texture2.c" )
 		tweakedNetwork = tweaks["out"].attributes( "/plane" )["surface"]
 		self.assertEqual( len( tweakedNetwork ), 4 )
-		self.assertEqual( tweakedNetwork.input( ( "surface", "c" ) ), ( "texture1", "out" ) )
-		self.assertEqual( tweakedNetwork.input( ( "texture1", "c" ) ), ( "texture2", "out" ) )
-		self.assertEqual( tweakedNetwork.input( ( "texture2", "c" ) ), ( "tweakShader", "out" ) )
+		self.assertEqual( tweakedNetwork.input( ( "surface", "c" ) ), ( "texture1", "c" ) )
+		self.assertEqual( tweakedNetwork.input( ( "texture1", "c" ) ), ( "texture2", "c" ) )
+		self.assertEqual( tweakedNetwork.input( ( "texture2", "c" ) ), ( "tweakShader", "c" ) )
 		self.assertEqual( tweakedNetwork.input( ( "tweakShader", "c" ) ), ( "", "" ) )
 
 		# Test using an auto-proxy on a parameter with no input ( it should apply the value to what the
@@ -124,16 +124,16 @@ class ShaderTweakProxyTest( GafferSceneTest.SceneTestCase ) :
 		tweaks["tweaks"][0]["name"].setValue( "c" )
 
 		specificProxy = GafferScene.ShaderTweakProxy()
-		specificProxy.loadShader( "test:testShader" )
+		specificProxy.loadShader( "test:simpleShader" )
 
 		specificProxy["parameters"]["targetShader"].setValue( "texture2" )
 
-		tweakShader["parameters"]["c"].setInput( specificProxy["out"]["out"] )
+		tweakShader["parameters"]["c"].setInput( specificProxy["out"]["c"] )
 
 		tweakedNetwork = tweaks["out"].attributes( "/plane" )["surface"]
 		self.assertEqual( len( tweakedNetwork ), 3 )
-		self.assertEqual( tweakedNetwork.input( ( "surface", "c" ) ), ( "tweakShader", "out" ) )
-		self.assertEqual( tweakedNetwork.input( ( "tweakShader", "c" ) ), ( "texture2", "out" ) )
+		self.assertEqual( tweakedNetwork.input( ( "surface", "c" ) ), ( "tweakShader", "c" ) )
+		self.assertEqual( tweakedNetwork.input( ( "tweakShader", "c" ) ), ( "texture2", "c" ) )
 
 		# Test error if we try to make a cycle
 		tweaks["tweaks"][0]["name"].setValue( "texture2.c" )
@@ -165,7 +165,7 @@ class ShaderTweakProxyTest( GafferSceneTest.SceneTestCase ) :
 		tweaks["shader"].setValue( "surface" )
 
 		tweaks["tweaks"].addChild( Gaffer.TweakPlug( "i", Gaffer.IntPlug() ) )
-		tweaks["tweaks"][0]["value"].setInput( tweakShader["out"]["r"] )
+		tweaks["tweaks"][0]["value"].setInput( tweakShader["out"]["c"]["r"] )
 
 		autoProxy = GafferScene.ShaderTweakProxy()
 		autoProxy.setupAutoProxy( Gaffer.IntPlug() )